Fix for #6 (sftp does not work)

+++ .bashrc
@@ -1,14 +1,17 @@
-wget --quiet -T 2 https://raw.githubusercontent.com/inferont/shell-scripts/master/.bashrc -O ~/.bashrc_new
-wget --quiet -T 2 https://raw.githubusercontent.com/inferont/shell-scripts/master/.commonrc -O ~/.commonrc_new
-if [ $(stat --printf="%s" ~/.bashrc_new) -gt 200 ]; then
-    rm ~/.bashrc
-    mv ~/.bashrc_new ~/.bashrc
+## If in tty download latest rc files and load
+if [ -t 1 ]; then
+    wget --quiet -T 2 https://raw.githubusercontent.com/inferont/shell-scripts/master/.bashrc -O ~/.bashrc_new
+    wget --quiet -T 2 https://raw.githubusercontent.com/inferont/shell-scripts/master/.commonrc -O ~/.commonrc_new
+    if [ $(stat --printf="%s" ~/.bashrc_new) -gt 200 ]; then
+        rm ~/.bashrc
+        mv ~/.bashrc_new ~/.bashrc
+    fi
+    if [ $(stat --printf="%s" ~/.commonrc_new) -gt 200 ]; then
+        rm ~/.commonrc
+        mv ~/.commonrc_new ~/.commonrc
+    fi
+    source ~/.commonrc
 fi
-if [ $(stat --printf="%s" ~/.commonrc_new) -gt 200 ]; then
-    rm ~/.commonrc
-    mv ~/.commonrc_new ~/.commonrc
-fi
-source ~/.commonrc
 if [ ! -f ~/bin/sshrc ]; then
     wget --quiet https://raw.githubusercontent.com/inferont/sshrc/master/sshrc -O ~/bin/sshrc
     chmod +x ~/bin/sshrc
